refactor: ImageDiffUtilCallback 패키지 변경
