+++ app/src/main/java/com/app/replace/ui/diaryeditor/adapter/DiaryEditorImageAdapter.kt
@@ -2,6 +2,7 @@ package com.app.replace.ui.diaryeditor.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
+import com.app.replace.ui.common.ImageDiffUtilCallback
 import com.app.replace.ui.diaryeditor.viewholder.DiaryEditorImageViewHolder
 
 class DiaryEditorImageAdapter(
